Fix code generation outputting call instead of get_local
This is sufficient that the generic list sum function now generates
correctly, which is sweet.

Basically the fix is to pass a set of names that are locals down the
generation process, and use that information to determine if we should
be making calls or getting locals.

Bit of a hack but it solves the immediate problem