Add code generation for closures
Previously, if a function closed over a variable, like when it's
declared inside of a let, the typechecker would let it pass, but the
code generator would not generate a correct implementation.

WebAssembly has no support for nested function declarations, so we have
to hoist declarations. In order to support closures, we now track which
variables have been closed over, and then add parameters to the hoisted
functions for closure arguments.

Identifiers also carry the closure binding information, to support
providiing the right locals at call time.